ENH: Remove unnecessary workflow steps
Remove unnecessary workflow steps:
- Remove the unnecessary `minimal` version step: no required packages
have a greater than or equal to version specification. This also avoids
the weak substitution of the sign to make the workflow fail because a
different Python version is being required, i.e.:
```
ERROR: Package 'braintransform' requires a different Python: 3.8.12 not in '==3.8'
```
in
https://github.com/brainhackorg/braintransform/runs/5175935193?check_suite_focus=true#step:7:58
- Remove caching the pip installation: this might be useful when the
number of required packages is large. It also avoids needing to specify
a file for the hash (`requirements.txt` no longer exists).